Added Homepage and picture
Added a homepage for activity #1. I'm a bit rusty with my HTML so I templated the sample code provided and plan on editing it to customize the webpage as I refreshen my knowledge.
I also uploaded a picture of myself and my kids to go with the site for a planned secondary page.